refactor: Make Ordinal docs-generation only
Drop the live-capture path for Ordinal (setup/capture/API client/event
triggers) and keep only the docs generator. Ordinal has no test-event
trigger and most events can't be driven through the API unattended, so the
generator is the practical way to keep the full sample set current.

With Ordinal no longer reusing the shared capture harness, that harness had
a single consumer (Scrapfly) and was no longer earned generality, so revert
it: restore scripts/scrapfly/ to its original self-contained form and
remove scripts/lib/. docs.ts is now self-contained (no shared imports).

Net change vs main is now just the Ordinal provider samples, the docs
generator, and a generate:ordinal script; Scrapfly is untouched.
