+++ package.json
@@ -16,9 +16,7 @@
     "dev:receiver": "nodemon --exec ./node_modules/.bin/ts-node requestReceiver.ts --watch *.ts",
     "setup:scrapfly": "./node_modules/.bin/ts-node scripts/scrapfly/setup.ts",
     "capture:scrapfly": "./node_modules/.bin/ts-node scripts/scrapfly/capture.ts",
-    "generate:ordinal": "./node_modules/.bin/ts-node scripts/ordinal/docs.ts",
-    "setup:ordinal": "./node_modules/.bin/ts-node scripts/ordinal/setup.ts",
-    "capture:ordinal": "./node_modules/.bin/ts-node scripts/ordinal/capture.ts"
+    "generate:ordinal": "./node_modules/.bin/ts-node scripts/ordinal/docs.ts"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
