+++ providers/ordinal/README.md
@@ -6,8 +6,8 @@ social media planning tool. Ordinal's webhook feature
 shipped 2026-04-07 and covers the full post lifecycle, social-profile
 connection state, approvals, and workspace invites.
 
-See [`scripts/ordinal/`](../../scripts/ordinal/) for the automation that
-provisions the capture pipeline and refreshes these payloads.
+See [`scripts/ordinal/`](../../scripts/ordinal/) for the generator that
+produces these payloads from Ordinal's docs.
 
 ## Delivery format
 
@@ -23,20 +23,22 @@ Every event shares one envelope:
 - `data` — event-specific; schema varies by `type`.
 - `createdAt` — ISO 8601 emission time.
 
-Ordinal does not document a webhook signature/HMAC scheme. The Hookdeck
-source is therefore a generic `WEBHOOK` type and the unguessable source
-URL acts as the shared secret. Custom headers can be attached per
-subscription via the webhook's `headers` field if downstream
+Ordinal does not document a webhook signature/HMAC scheme; the unguessable
+delivery URL acts as the shared secret, and custom headers can be attached
+per subscription via the webhook's `headers` field if downstream
 verification is needed.
 
 ## ⚠️ Provenance of these payloads
 
 > **All payloads in `latest/` are docs-sourced and generated by a script.**
-> Ordinal has ~20 event types and no test-event trigger, so rather than
-> hand-write payloads, [`scripts/ordinal/docs.ts`](../../scripts/ordinal/docs.ts)
-> scrapes Ordinal's documentation: it discovers every event page from the
-> docs index (`llms.txt`) and extracts the canonical example payload each
-> page publishes in a `json` fence, then writes it here in this repo's
+> Ordinal has ~20 event types and no test-event trigger, and most events
+> can't be produced through the API unattended (OAuth connects, real
+> publishes, approver/invitee actions, the debounced in-app editor), so a
+> live capture of the full set isn't practical. Instead,
+> [`scripts/ordinal/docs.ts`](../../scripts/ordinal/docs.ts) scrapes
+> Ordinal's documentation: it discovers every event page from the docs
+> index (`llms.txt`) and extracts the canonical example payload each page
+> publishes in a `json` fence, then writes it here in this repo's
 > `{ headers, body, topic }` shape. Regenerate any time with:
 >
 > ```sh
@@ -48,41 +50,30 @@ verification is needed.
 > mistaken for real account traffic, and `headers` are representative, not
 > recorded (the docs don't enumerate delivery headers).
 
-A subset can additionally be refreshed with **real** deliveries by running
-`yarn capture:ordinal` against a live Ordinal workspace (see
-[`scripts/ordinal/`](../../scripts/ordinal/)) — this performs the matching
-API operation and records the actual webhook. The rest require a side
-channel the API can't drive unattended and stay docs-sourced. The table
-below tracks which is which.
-
 ## Event taxonomy
 
-| `type` | Group | Capture status | Why |
-|--------|-------|----------------|-----|
-| `post.created` | Post | **Live-capturable** | `POST /posts` |
-| `post.scheduled` | Post | **Live-capturable** | `POST /posts/{id}/schedule` |
-| `post.rescheduled` | Post | **Live-capturable** | re-schedule with a new time |
-| `post.unscheduled` | Post | **Live-capturable** | `POST /posts/{id}/unschedule` |
-| `post.archived` | Post | **Live-capturable** | `POST /posts/{id}/archive` |
-| `invite.created` | Invite | **Live-capturable** | `POST /invites` |
-| `post.published` | Post | Docs-sourced | needs a connected channel + a real publish to the platform |
-| `post.publish_failed` | Post | Docs-sourced | needs a real publish attempt to fail |
-| `post.permanently_deleted` | Post | Docs-sourced | no API endpoint; trash-retention or in-app delete |
-| `post.content.edited` | Post | Docs-sourced | in-app editor only, ~5 min debounce |
-| `post.comment.created` | Post | Docs-sourced | no comment API |
-| `post.inline_comment.created` | Post | Docs-sourced | no comment API |
-| `post.approval.requested` | Approval | Docs-sourced | needs another workspace member as approver |
-| `post.approval.approved` | Approval | Docs-sourced | needs an approver to grant approval |
-| `campaign.approval.requested` | Approval | Docs-sourced | campaign approvals are in-app |
-| `campaign.approval.approved` | Approval | Docs-sourced | needs an approver to grant approval |
-| `invite.accepted` | Invite | Docs-sourced | needs the invitee to accept |
-| `social_profile.connected` | Social profile | Docs-sourced | OAuth connect flow |
-| `social_profile.disconnected` | Social profile | Docs-sourced | OAuth/in-app disconnect |
-| `social_profile.reconnect_needed` | Social profile | Docs-sourced | emitted when a token expires |
-
-> `post.approval.requested` is API-reachable in principle
-> (`POST /approvals`) but needs a second member in the workspace to
-> request approval from, so it isn't driven by the unattended capture.
+| `type` | Group |
+|--------|-------|
+| `post.created` | Post |
+| `post.scheduled` | Post |
+| `post.rescheduled` | Post |
+| `post.unscheduled` | Post |
+| `post.published` | Post |
+| `post.publish_failed` | Post |
+| `post.archived` | Post |
+| `post.permanently_deleted` | Post |
+| `post.content.edited` | Post |
+| `post.comment.created` | Post |
+| `post.inline_comment.created` | Post |
+| `post.approval.requested` | Approval |
+| `post.approval.approved` | Approval |
+| `campaign.approval.requested` | Approval |
+| `campaign.approval.approved` | Approval |
+| `invite.created` | Invite |
+| `invite.accepted` | Invite |
+| `social_profile.connected` | Social profile |
+| `social_profile.disconnected` | Social profile |
+| `social_profile.reconnect_needed` | Social profile |
 
 ### Field notes worth knowing
 
@@ -95,16 +86,12 @@ below tracks which is which.
 - The three `social_profile.*` events share one `profile` sub-schema,
   differing only in the actor/timestamp fields.
 
-## Regenerating / re-capturing
+## Regenerating
 
 ```sh
-yarn generate:ordinal  # rebuild ALL payloads from Ordinal's published docs (no creds)
-# — or, for real deliveries from a live workspace —
-yarn setup:ordinal     # one-time: Hookdeck source + Ordinal webhook subscription
-yarn capture:ordinal   # overwrite the live-capturable subset with real captures
+yarn generate:ordinal
 ```
 
-Full instructions, prerequisites, and design notes are in
-[`scripts/ordinal/README.md`](../../scripts/ordinal/README.md). Review any
-newly captured files before committing — real deliveries carry your test
-workspace's slug, user emails, and post URLs.
+Rebuilds every payload from Ordinal's published docs. No credentials
+needed. See [`scripts/ordinal/README.md`](../../scripts/ordinal/README.md)
+for how it works and how to handle the docs adding or removing an event.
