--- scripts/lib/index.ts
@@ -1,9 +1,0 @@
-// Provider-agnostic harness toolkit. Each provider directory
-// (scripts/<provider>/) composes these helpers with a small,
-// provider-specific config + driver. Nothing here knows about a
-// specific webhook source.
-export * from "./env";
-export * from "./prompt";
-export * from "./process";
-export * from "./hookdeck";
-export * from "./capture";
