--- scripts/ordinal/.env.example
@@ -1,21 +1,0 @@
-# Copy to scripts/ordinal/.env.local (preferred) or .env.local at the
-# repo root, and fill in. Either path is gitignored.
-
-# Hookdeck project API key — https://dashboard.hookdeck.com/settings/project/secrets
-HOOKDECK_API_KEY=
-
-# Ordinal API key (workspace-scoped, Pro plan or higher) —
-# https://app.tryordinal.com/settings/integrations/api  (format: ord_...)
-ORDINAL_API_KEY=
-
-# The Hookdeck source name (idempotently created by `yarn setup:ordinal`).
-HOOKDECK_SOURCE_NAME=ordinal
-
-# The name of the Ordinal webhook subscription created by setup.
-ORDINAL_WEBHOOK_NAME=samples-capture
-
-# Written by `yarn setup:ordinal` — the id of the upserted Ordinal webhook.
-ORDINAL_WEBHOOK_ID=
-
-# Optional override for the Ordinal API base URL.
-ORDINAL_API_BASE_URL=https://app.tryordinal.com/api/v1
