+++ scripts/scrapfly/setup.ts
@@ -4,12 +4,8 @@ import {
   requireEnv,
   prompt,
   promptHidden,
-  hookdeckCi,
-  upsertSource,
+  runHookdeck,
 } from "./lib";
-
-const SOURCE_DESCRIPTION =
-  "Captures sample webhooks from Scrapfly for the webhook-samples repo";
 
 async function main() {
   console.log("Scrapfly capture — setup\n");
@@ -20,16 +16,54 @@ async function main() {
   const webhookName = env.SCRAPFLY_WEBHOOK_NAME || "samples-capture";
 
   console.log("1/4 Authenticating Hookdeck CLI for this project...");
-  hookdeckCi(HOOKDECK_API_KEY, "scrapfly-samples");
-
-  console.log(
-    `\n2/4 Upserting Hookdeck source "${sourceName}" (initial, no secret)...`
+  const ci = runHookdeck(
+    ["ci", "--api-key", HOOKDECK_API_KEY, "--local", "--name", "scrapfly-samples"],
+    { capture: true }
   );
-  const sourceUrl = upsertSource({
-    name: sourceName,
-    type: "WEBHOOK",
-    description: SOURCE_DESCRIPTION,
-  });
+  if (ci.code !== 0) {
+    console.error("hookdeck ci failed:");
+    console.error(ci.stderr || ci.stdout);
+    process.exit(1);
+  }
+  process.stdout.write(ci.stdout);
+
+  console.log(`\n2/4 Upserting Hookdeck source "${sourceName}" (initial, no secret)...`);
+  const initialUpsert = runHookdeck(
+    [
+      "gateway",
+      "source",
+      "upsert",
+      sourceName,
+      "--type",
+      "WEBHOOK",
+      "--description",
+      "Captures sample webhooks from Scrapfly for the webhook-samples repo",
+      "--output",
+      "json",
+    ],
+    { capture: true }
+  );
+  if (initialUpsert.code !== 0) {
+    console.error("hookdeck gateway source upsert failed:");
+    console.error(initialUpsert.stderr || initialUpsert.stdout);
+    process.exit(1);
+  }
+
+  let sourceUrl: string | undefined;
+  try {
+    const parsed = JSON.parse(initialUpsert.stdout);
+    sourceUrl = parsed?.url || parsed?.source?.url;
+  } catch (e) {
+    console.error("Could not parse upsert output as JSON:");
+    console.error(initialUpsert.stdout);
+    process.exit(1);
+  }
+  if (!sourceUrl) {
+    console.error("No source URL returned from Hookdeck. Raw output:");
+    console.error(initialUpsert.stdout);
+    process.exit(1);
+  }
+
   console.log(`   Source URL: ${sourceUrl}`);
 
   console.log(`\n3/4 Configure Scrapfly manually:
@@ -71,12 +105,32 @@ async function main() {
   console.log(
     `\n4/4 Re-upserting source as type SCRAPFLY with signature verification...`
   );
-  upsertSource({
-    name: sourceName,
-    type: "SCRAPFLY",
-    description: SOURCE_DESCRIPTION,
-    webhookSecret,
-  });
+  const secureUpsert = runHookdeck(
+    [
+      "gateway",
+      "source",
+      "upsert",
+      sourceName,
+      "--type",
+      "SCRAPFLY",
+      "--webhook-secret",
+      webhookSecret,
+      "--description",
+      "Captures sample webhooks from Scrapfly for the webhook-samples repo",
+      "--output",
+      "json",
+    ],
+    { capture: true }
+  );
+  if (secureUpsert.code !== 0) {
+    console.error("Failed to upgrade source to SCRAPFLY type:");
+    console.error(secureUpsert.stderr || secureUpsert.stdout);
+    console.error(
+      "\nFallback: source remains as type WEBHOOK without verification. " +
+        "Re-run `yarn setup:scrapfly` to retry."
+    );
+    process.exit(1);
+  }
   console.log(`   Source verification enabled.`);
 
   console.log(`\nDone. Next: ensure SCRAPFLY_API_KEY is set in .env.local, then run:
